requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart
PyPDF2>=3.0.0
pinecone-client>=4.0.0
google-generativeai>=0.3.0
python-dotenv
pydantic>=2.0.0